refactor(git): extract task patch engine
